mod_rayo: Translate XMPP messages to SIP messages
Call-bound XMPP messages are translated to SIP messages via
SWITCH_EVENT_SEND_MESSAGE in a similar manner to that described in
draft-ietf-stox-im-06. Messages with a type of "normal" are directed
to the caller. Other types receive a feature-not-implemented response
but it is envisaged that the "groupchat" type could be used to direct
the message to all joined parties.